requirements.txt
tensorflow
streamlit
pillow 
cvzone
numpy
